Prepare mzdb4s for release v0.4.3: * add support for FAIMS Thermo RAW files conversion (new --split-faims parameters, which generates multiple mzDB file, one per CV value) * add support for native builds on Windows (depends on SN 0.4.1, build static&shared native libraries for Windows, add strptime custom C++ library) * fix the parsing of the ActivationType information for Thermo RAW files * fix spectrum "initial ID" issue (use always spectrum ID instead, since it Proline is currently crashing if IDs are not identical), this should avoid issues related to empty spectra * fix spectra Retention Time order issue for TimsTOF data * update the ThermoRawFileParser to be able to read sample information from raw files * improve performance of AbstractSpectrumHeaderReader * improve performance of MgfSpectrumSerializer (use bytes instead of string representation) * update dependencies (plugins, SBT, and project dependencies) to latest versions

+++ io-thermo/native/src/main/scala/ThermoToMzDB.scala
@@ -23,8 +23,8 @@ object ThermoToMzDB {
   implicit val sf: ISQLiteFactory = SQLiteFactory
 
   private def readLink(link: CString)(implicit z: Zone): CString = {
-    val buffer: CString = alloc[Byte](limits.PATH_MAX.toUInt)
-    readlink(link, buffer, (limits.PATH_MAX - 1).toUInt) match {
+    val buffer: CString = alloc[Byte](limits.PATH_MAX)
+    readlink(link, buffer, limits.PATH_MAX - 1.toUInt) match {
       case -1 =>
         null
       case read =>
